Add a camera which don't work very well

+++ src/Mot/AnimatedSprite.cpp
@@ -63,7 +63,7 @@ int AnimatedSprite::update(double seconds)
 	return 0;
 }
 
-int AnimatedSprite::draw(Vector2d pos, float scale)
+int AnimatedSprite::draw(Vector2d pos, double scale)
 {
     if (_tempo > (1./_framePerSecs[_actualSubRect.x])) {
         _actualSubRect.y++;
